requirement.txt
asgiref==3.6.0
Django==4.1.4
django-environ==0.9.0
django-hashid-field==3.3.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
hashids==1.3.1
pytz==2022.7
sqlparse==0.4.3
